+++ src/components/Label/index.styles.js
@@ -1,0 +1,27 @@
+import styled from 'styled-components';
+
+export const LabelWrapper = styled.li`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 20px 15px;
+  border-bottom: 1px solid #ccc;
+`;
+
+export const LabelBlock = styled.div`
+  flex: 0.5;
+`;
+
+export const Desc = styled.div`
+  flex: 1;
+`;
+
+export const Edit = styled.button`
+  padding: 10px;
+  cursor: pointer;
+`;
+
+export const Delete = styled.button`
+  padding: 10px;
+  cursor: pointer;
+`;
