+++ src/components/Menu/MenuTab.js
@@ -1,5 +1,14 @@
 import styled from 'styled-components';
 
+const MenuTab = () => {
+  return (
+    <MenuTabList>
+      <li>Labels</li>
+      <li>Milestones</li>
+    </MenuTabList>
+  );
+};
+
 const MenuTabList = styled.ul`
   display: flex;
 
@@ -10,14 +19,4 @@ const MenuTabList = styled.ul`
     cursor: pointer;
   }
 `;
-
-const MenuTab = () => {
-  return (
-    <MenuTabList>
-      <li>Labels</li>
-      <li>Milestones</li>
-    </MenuTabList>
-  );
-};
-
 export default MenuTab;
