+++ src/components/Menu/NewLabelButton.js
@@ -1,13 +1,13 @@
 import styled from 'styled-components';
 
-const Button = styled.button`
-  cursor: pointer;
-`;
-
 const NewLabelButton = ({ newLabel, setNewLabel }) => {
   const openNewLabel = () => setNewLabel(!newLabel);
 
   return <Button onClick={openNewLabel}>New label</Button>;
 };
+
+const Button = styled.button`
+  cursor: pointer;
+`;
 
 export default NewLabelButton;
