+++ src/components/NewLabel/index.styles.js
@@ -1,0 +1,32 @@
+import styled from 'styled-components';
+
+export const NewLabelBlock = styled.div`
+  display: flex;
+  flex-direction: column;
+  background-color: #888;
+  padding: 10px;
+`;
+
+export const LabelPreviewBlock = styled.div`
+  padding: 10px 0;
+`;
+
+export const NewLabelContainer = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-end;
+  margin-top: 15px;
+`;
+
+export const DescWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+`;
+
+export const ColorInputWrapper = styled.div``;
+
+export const LabelWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+`;
